+++ frontend/src/app/features/docente/dashboard/docente-dashboard.component.html
@@ -1,5 +1,128 @@
 <h2 class="ttl">Panel de docente</h2>
 
+<div class="card solicitudes-card">
+  <div class="card-header">
+    <div class="card-title">
+      <h3>Solicitudes de reunión</h3>
+      <p class="muted">Revisa las solicitudes de tus alumnos y agenda directamente</p>
+    </div>
+    <span class="chip" *ngIf="solicitudesPendientes.length">
+      {{ solicitudesPendientes.length }} pendiente{{ solicitudesPendientes.length === 1 ? '' : 's' }}
+    </span>
+  </div>
+
+  <div class="alert success" *ngIf="solicitudesMensaje">{{ solicitudesMensaje }}</div>
+  <div class="alert error" *ngIf="solicitudesError">{{ solicitudesError }}</div>
+
+  <p class="muted" *ngIf="solicitudesCargando">Cargando solicitudes…</p>
+  <p class="muted" *ngIf="!solicitudesCargando && !solicitudes.length">No hay solicitudes registradas.</p>
+
+  <div class="solicitud" *ngFor="let solicitud of solicitudes">
+    <header>
+      <div>
+        <h4>{{ solicitud.alumno?.nombre || 'Alumno' }}</h4>
+        <small class="muted">Enviada el {{ fechaHora(solicitud.creadoEn) }}</small>
+      </div>
+      <span
+        class="estado"
+        [class.pendiente]="solicitud.estado === 'pendiente'"
+        [class.aprobada]="solicitud.estado === 'aprobada'"
+        [class.rechazada]="solicitud.estado === 'rechazada'"
+      >
+        {{ estadoSolicitudLabel(solicitud.estado) }}
+      </span>
+    </header>
+
+    <p class="motivo">{{ solicitud.motivo }}</p>
+    <p class="muted" *ngIf="solicitud.disponibilidadSugerida">
+      Disponibilidad sugerida: <strong>{{ solicitud.disponibilidadSugerida }}</strong>
+    </p>
+
+    <div class="acciones" *ngIf="solicitud.estado === 'pendiente'">
+      <button class="btn-edit" type="button" (click)="abrirAprobacion(solicitud)" [disabled]="procesandoSolicitud">
+        Aprobar y agendar
+      </button>
+      <button class="btn-cancel" type="button" (click)="abrirRechazo(solicitud)" [disabled]="procesandoSolicitud">
+        Rechazar
+      </button>
+    </div>
+
+    <form
+      class="panel-form"
+      *ngIf="seleccionada?.id === solicitud.id && modo === 'aprobar'"
+      [formGroup]="aprobarForm"
+      (ngSubmit)="confirmarAprobacion()"
+    >
+      <h4>Agendar reunión</h4>
+      <div class="grid form-grid">
+        <label>
+          Fecha
+          <input type="date" formControlName="fecha" />
+        </label>
+        <label>
+          Hora inicio
+          <input type="time" formControlName="horaInicio" />
+        </label>
+        <label>
+          Hora término
+          <input type="time" formControlName="horaTermino" />
+        </label>
+        <label>
+          Modalidad
+          <select formControlName="modalidad">
+            <option value="presencial">Presencial</option>
+            <option value="online">Online</option>
+          </select>
+        </label>
+      </div>
+      <label class="block">
+        Comentario (opcional)
+        <textarea rows="3" formControlName="comentario" placeholder="Instrucciones para el estudiante"></textarea>
+      </label>
+      <div class="acciones">
+        <button class="btn-edit" type="submit" [disabled]="procesandoSolicitud">
+          {{ procesandoSolicitud ? 'Guardando…' : 'Confirmar reunión' }}
+        </button>
+        <button class="btn-session" type="button" (click)="cancelarAccion()">Cancelar</button>
+      </div>
+    </form>
+
+    <form
+      class="panel-form"
+      *ngIf="seleccionada?.id === solicitud.id && modo === 'rechazar'"
+      [formGroup]="rechazoForm"
+      (ngSubmit)="confirmarRechazo()"
+    >
+      <h4>Rechazar solicitud</h4>
+      <label class="block">
+        Comentario (opcional)
+        <textarea rows="3" formControlName="comentario" placeholder="Explica al estudiante el motivo"></textarea>
+      </label>
+      <div class="acciones">
+        <button class="btn-cancel" type="submit" [disabled]="procesandoSolicitud">
+          {{ procesandoSolicitud ? 'Guardando…' : 'Confirmar rechazo' }}
+        </button>
+        <button class="btn-session" type="button" (click)="cancelarAccion()">Cancelar</button>
+      </div>
+    </form>
+  </div>
+
+  <div class="historial" *ngIf="solicitudesResueltas.length">
+    <h4>Historial de solicitudes</h4>
+    <ul class="historial-lista">
+      <li *ngFor="let solicitud of solicitudesResueltas">
+        <div>
+          <strong>{{ solicitud.alumno?.nombre || 'Alumno' }}</strong>
+          <p class="muted">Actualizada el {{ fechaHora(solicitud.actualizadoEn) }}</p>
+        </div>
+        <span class="estado" [class.aprobada]="solicitud.estado === 'aprobada'" [class.rechazada]="solicitud.estado === 'rechazada'">
+          {{ estadoSolicitudLabel(solicitud.estado) }}
+        </span>
+      </li>
+    </ul>
+  </div>
+</div>
+
 <h3 class="section-subtitle">Estudiantes Asignados</h3>
 
 <div class="filters">
